Sane cursor addition and deletion
Old behaviour:

On branch the new cursor is added at the end of the array, current cursor doesn't change.
On delete the current cursor is deleted, current cursor becomes the cursor after the next cursor.

New behaviour:

On branch the new cursor is added after the current cursor, and becomes the current cursor.
On delete the current cursor is deleted, and the previous cursor becomes the current cursor.